More fixes for new tree

+++ Packages/Uintah/Core/Datatypes/CCVectorField.cc
@@ -24,7 +24,7 @@ CCVectorField::CCVectorField(const CCVectorField& copy)
   :VectorFieldRG( copy ), grid(copy.grid), level(copy.level),
     _varname(copy._varname), _matIndex(copy._matIndex)
 {
-  for(int i = 0; i < copy._vars.size(); i++){
+  for(int i = 0; i < (int)copy._vars.size(); i++){
     _vars.push_back( copy._vars[i] );
   }
 }
@@ -36,7 +36,7 @@ CCVectorField::CCVectorField(GridP grid, LevelP level,
   : VectorFieldRG(), grid(grid), level(level),
     _varname(var), _matIndex(mat)
 {
-  for(int i = 0; i < vars.size(); i++){
+  for(int i = 0; i < (int)vars.size(); i++){
     _vars.push_back( vars[i]);
   }
 }
@@ -110,8 +110,7 @@ int CCVectorField::interpolate(const Point& p, Vector& value, int&,
 
 int CCVectorField::interpolate(const Point& p, Vector& value)
 {
-using namespace SCIRun;
-
+  using namespace SCIRun;
 
   int i;
   IntVector index;
@@ -124,7 +123,7 @@ using namespace SCIRun;
     }
   }
 
-  if (i >= _vars.size() || r == level->patchesEnd() )
+  if (i >= (int)_vars.size() || r == level->patchesEnd() )
     return 0;
   
   value = _vars[i][index];
@@ -132,4 +131,3 @@ using namespace SCIRun;
 }
 
 } // End namespace Uintah
-
